Split command line options validator
Copilot-Session: 28a10aca-f4aa-4519-a92b-c729313d2968
