+++ src/Platform/Microsoft.Testing.Platform/CommandLine/CommandLineOptionsValidator.cs
@@ -9,7 +9,7 @@ using Microsoft.Testing.Platform.Tools;
 
 namespace Microsoft.Testing.Platform.CommandLine;
 
-internal static class CommandLineOptionsValidator
+internal static partial class CommandLineOptionsValidator
 {
     // Options whose values are read directly from CommandLineParseResult during platform
     // bootstrap, before IConfiguration (and therefore testconfig.json) is built. Setting them
@@ -118,415 +118,5 @@ internal static class CommandLineOptionsValidator
             : ValidateExtensionOptionsDoNotContainReservedOptions(toolOptionsByProvider, systemOptionsByProvider) is { IsValid: false } reservedOptionResult
                 ? reservedOptionResult
                 : ValidateOptionsAreNotDuplicated(toolOptionsByProvider, systemOptionsByProvider);
-    }
-
-    private static ValidationResult ValidateExtensionOptionsDoNotContainReservedPrefix(
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> extensionOptionsByProvider)
-    {
-        StringBuilder? stringBuilder = null;
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> providerWithOptions in extensionOptionsByProvider)
-        {
-            foreach (CommandLineOption option in providerWithOptions.Value)
-            {
-                if (option.IsBuiltIn)
-                {
-                    continue;
-                }
-
-                string trimmedOption = option.Name.Trim(CommandLineParseResult.OptionPrefix);
-                if (trimmedOption.StartsWith("internal", StringComparison.OrdinalIgnoreCase)
-                    || option.Name.StartsWith("-internal", StringComparison.OrdinalIgnoreCase))
-                {
-                    stringBuilder ??= new();
-                    ICommandLineOptionsProvider commandLineOptionsProvider = providerWithOptions.Key;
-                    stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionIsUsingReservedPrefix, trimmedOption, commandLineOptionsProvider.DisplayName, commandLineOptionsProvider.Uid));
-                }
-            }
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static ValidationResult ValidateExtensionOptionsDoNotContainReservedOptions(
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> extensionOptionsByProvider,
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> systemOptionsByProvider)
-    {
-        // Create a HashSet of all system option names for faster lookup. OrdinalIgnoreCase keeps
-        // duplicate detection consistent with the case-insensitive lookups performed later in the
-        // pipeline and with how testconfig.json stores keys.
-        var systemOptionNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> provider in systemOptionsByProvider)
-        {
-            foreach (CommandLineOption option in provider.Value)
-            {
-                systemOptionNames.Add(option.Name);
-            }
-        }
-
-        // Aggregate reserved options by name and track all offending providers
-        var reservedOptionToProviderNames = new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase);
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> kvp in extensionOptionsByProvider)
-        {
-            foreach (CommandLineOption option in kvp.Value)
-            {
-                if (systemOptionNames.Contains(option.Name))
-                {
-                    if (!reservedOptionToProviderNames.TryGetValue(option.Name, out HashSet<string>? providerNames))
-                    {
-                        providerNames = [];
-                        reservedOptionToProviderNames[option.Name] = providerNames;
-                    }
-
-                    providerNames.Add(kvp.Key.DisplayName);
-                }
-            }
-        }
-
-        StringBuilder? stringBuilder = null;
-        foreach (KeyValuePair<string, HashSet<string>> kvp in reservedOptionToProviderNames)
-        {
-            stringBuilder ??= new StringBuilder();
-            stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionIsReserved, kvp.Key, string.Join("', '", kvp.Value)));
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static ValidationResult ValidateOptionsAreNotDuplicated(
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> extensionOptionsByProvider,
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> systemOptionsByProvider)
-    {
-        // Use a dictionary to track option names and their distinct providers. OrdinalIgnoreCase
-        // ensures we catch case-differing duplicates (e.g. "Timeout" vs "timeout") with a friendly
-        // error rather than a later ArgumentException when building the lookup dictionary.
-        // We cover both extension and system providers so that any pair of case-differing
-        // duplicates — extension/extension, extension/system, or system/system — surfaces as a
-        // ValidationResult.Invalid rather than crashing the platform.
-        var optionNameToProviders = new Dictionary<string, HashSet<ICommandLineOptionsProvider>>(StringComparer.OrdinalIgnoreCase);
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> kvp in extensionOptionsByProvider.Concat(systemOptionsByProvider))
-        {
-            ICommandLineOptionsProvider provider = kvp.Key;
-            foreach (CommandLineOption option in kvp.Value)
-            {
-                string name = option.Name;
-                if (!optionNameToProviders.TryGetValue(name, out HashSet<ICommandLineOptionsProvider>? providers))
-                {
-                    providers = [];
-                    optionNameToProviders[name] = providers;
-                }
-
-                providers.Add(provider);
-            }
-        }
-
-        // Check for duplications
-        StringBuilder? stringBuilder = null;
-        foreach (KeyValuePair<string, HashSet<ICommandLineOptionsProvider>> kvp in optionNameToProviders)
-        {
-            if (kvp.Value.Count > 1)
-            {
-                string duplicatedOption = kvp.Key;
-                stringBuilder ??= new();
-                IEnumerable<string> faultyProvidersDisplayNames = kvp.Value.Select(p => p.DisplayName);
-                stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionIsDeclaredByMultipleProviders, duplicatedOption, string.Join("', '", faultyProvidersDisplayNames)));
-            }
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static ValidationResult ValidateNoUnknownOptions(
-        CommandLineParseResult parseResult,
-        IReadOnlyList<JsonCommandLineOptionEntry>? jsonCommandLineOptions,
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> extensionOptionsByProvider,
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> systemOptionsByProvider)
-    {
-        // Use OrdinalIgnoreCase so a JSON entry like "Timeout" resolves to the registered "timeout"
-        // option (testconfig.json keys are case-insensitive everywhere else in the platform). CLI
-        // parsing is already case-sensitive but a case-insensitive lookup is a strict superset and
-        // does not change CLI behavior.
-        var validOptionNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> provider in extensionOptionsByProvider)
-        {
-            foreach (CommandLineOption option in provider.Value)
-            {
-                validOptionNames.Add(option.Name);
-            }
-        }
-
-        foreach (KeyValuePair<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>> provider in systemOptionsByProvider)
-        {
-            foreach (CommandLineOption option in provider.Value)
-            {
-                validOptionNames.Add(option.Name);
-            }
-        }
-
-        StringBuilder? stringBuilder = null;
-        foreach (CommandLineParseOption optionRecord in parseResult.Options)
-        {
-            if (!validOptionNames.Contains(optionRecord.Name))
-            {
-                stringBuilder ??= new();
-                stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineUnknownOption, optionRecord.Name));
-            }
-        }
-
-        // Also surface unknown entries under the testconfig.json "commandLineOptions" section.
-        // We intentionally validate even when the CLI provides a matching option of the same name
-        // (which would shadow the JSON value at lookup time): a JSON typo silently overridden by
-        // the CLI is still a typo that the user wants to know about.
-        if (jsonCommandLineOptions is { Count: > 0 })
-        {
-            foreach (JsonCommandLineOptionEntry entry in jsonCommandLineOptions)
-            {
-                if (!validOptionNames.Contains(entry.OptionName))
-                {
-                    stringBuilder ??= new();
-                    string innerError = string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineUnknownOption, entry.OptionName);
-                    stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.JsonCommandLineOptionsValidationErrorPrefix, innerError));
-                }
-            }
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static ValidationResult ValidateNoBootstrapOnlyOptionsInJson(
-        IReadOnlyList<JsonCommandLineOptionEntry>? jsonCommandLineOptions)
-    {
-        if (jsonCommandLineOptions is not { Count: > 0 })
-        {
-            return ValidationResult.Valid();
-        }
-
-        StringBuilder? stringBuilder = null;
-        foreach (JsonCommandLineOptionEntry entry in jsonCommandLineOptions)
-        {
-            if (!BootstrapOnlyOptions.Contains(entry.OptionName))
-            {
-                continue;
-            }
-
-            stringBuilder ??= new();
-            string innerError = string.Format(CultureInfo.InvariantCulture, PlatformResources.JsonCommandLineOptionIsBootstrapOnlyErrorMessage, entry.OptionName);
-            stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.JsonCommandLineOptionsValidationErrorPrefix, innerError));
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static ValidationResult ValidateOptionsArgumentArity(
-        CommandLineParseResult parseResult,
-        IReadOnlyList<JsonCommandLineOptionEntry>? jsonCommandLineOptions,
-        Dictionary<string, (ICommandLineOptionsProvider Provider, CommandLineOption Option)> providerAndOptionByOptionName)
-    {
-        StringBuilder? stringBuilder = null;
-        foreach (IGrouping<string, CommandLineParseOption> groupedOptions in parseResult.Options.GroupBy(x => x.Name, StringComparer.OrdinalIgnoreCase))
-        {
-            // getting the arguments count for an option.
-            int arity = 0;
-            foreach (CommandLineParseOption optionEntry in groupedOptions)
-            {
-                arity += optionEntry.Arguments.Length;
-            }
-
-            string optionName = groupedOptions.Key;
-            (ICommandLineOptionsProvider provider, CommandLineOption option) = providerAndOptionByOptionName[optionName];
-
-            AppendArityErrorIfNeeded(stringBuilder: ref stringBuilder, arity, optionName, provider, option, jsonPrefix: false);
-        }
-
-        // Apply the same arity rules to entries sourced from testconfig.json. We skip
-        // explicit disable entries ("foo": false) because they convey "the option is not set" —
-        // there are no arguments to validate. Unknown options were rejected earlier.
-        if (jsonCommandLineOptions is { Count: > 0 })
-        {
-            foreach (JsonCommandLineOptionEntry entry in jsonCommandLineOptions)
-            {
-                if (entry.IsDisabled)
-                {
-                    continue;
-                }
-
-                if (!providerAndOptionByOptionName.TryGetValue(entry.OptionName, out (ICommandLineOptionsProvider Provider, CommandLineOption Option) match))
-                {
-                    continue;
-                }
-
-                AppendArityErrorIfNeeded(stringBuilder: ref stringBuilder, entry.Arguments.Count, entry.OptionName, match.Provider, match.Option, jsonPrefix: true);
-            }
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static void AppendArityErrorIfNeeded(
-        ref StringBuilder? stringBuilder,
-        int arity,
-        string optionName,
-        ICommandLineOptionsProvider provider,
-        CommandLineOption option,
-        bool jsonPrefix)
-    {
-        string? message = null;
-        if (arity > option.Arity.Max && option.Arity.Max == 0)
-        {
-            message = string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionExpectsNoArguments, optionName, provider.DisplayName, provider.Uid);
-        }
-        else if (arity < option.Arity.Min)
-        {
-            message = string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionExpectsAtLeastArguments, optionName, provider.DisplayName, provider.Uid, option.Arity.Min);
-        }
-        else if (arity > option.Arity.Max)
-        {
-            message = string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineOptionExpectsAtMostArguments, optionName, provider.DisplayName, provider.Uid, option.Arity.Max);
-        }
-
-        if (message is null)
-        {
-            return;
-        }
-
-        stringBuilder ??= new();
-        stringBuilder.AppendLine(jsonPrefix
-            ? string.Format(CultureInfo.InvariantCulture, PlatformResources.JsonCommandLineOptionsValidationErrorPrefix, message)
-            : message);
-    }
-
-    private static async Task<ValidationResult> ValidateOptionsArgumentsAsync(
-        CommandLineParseResult parseResult,
-        IReadOnlyList<JsonCommandLineOptionEntry>? jsonCommandLineOptions,
-        Dictionary<string, (ICommandLineOptionsProvider Provider, CommandLineOption Option)> providerAndOptionByOptionName)
-    {
-        if (parseResult is null)
-        {
-            throw new ArgumentNullException(nameof(parseResult));
-        }
-
-        StringBuilder? stringBuilder = null;
-        foreach (IGrouping<string, CommandLineParseOption> optionRecords in parseResult.Options.GroupBy(
-            record => record.Name,
-            StringComparer.OrdinalIgnoreCase))
-        {
-            (ICommandLineOptionsProvider provider, CommandLineOption option) = providerAndOptionByOptionName[optionRecords.Key];
-            string[] arguments = [.. optionRecords.SelectMany(record => record.Arguments)];
-            ValidationResult result = await provider.ValidateOptionArgumentsAsync(option, arguments).ConfigureAwait(false);
-            if (!result.IsValid)
-            {
-                stringBuilder ??= new();
-                stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineInvalidArgumentsForOption, optionRecords.Key, result.ErrorMessage));
-            }
-        }
-
-        // Apply the per-option argument validators to JSON-sourced entries as well. Skip disabled
-        // entries (nothing to validate) and entries that the prior arity pass already flagged
-        // (calling a provider's validator with too-few/too-many arguments may produce confusing
-        // secondary errors or, worse, index out of bounds inside the validator itself).
-        if (jsonCommandLineOptions is { Count: > 0 })
-        {
-            foreach (JsonCommandLineOptionEntry entry in jsonCommandLineOptions)
-            {
-                if (entry.IsDisabled)
-                {
-                    continue;
-                }
-
-                if (!providerAndOptionByOptionName.TryGetValue(entry.OptionName, out (ICommandLineOptionsProvider Provider, CommandLineOption Option) match))
-                {
-                    continue;
-                }
-
-                if (entry.Arguments.Count < match.Option.Arity.Min || entry.Arguments.Count > match.Option.Arity.Max)
-                {
-                    continue;
-                }
-
-                string[] argumentsArray = entry.Arguments as string[] ?? entry.Arguments.ToArray();
-                ValidationResult result = await match.Provider.ValidateOptionArgumentsAsync(match.Option, argumentsArray).ConfigureAwait(false);
-                if (!result.IsValid)
-                {
-                    stringBuilder ??= new();
-                    string innerError = string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineInvalidArgumentsForOption, entry.OptionName, result.ErrorMessage);
-                    stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.JsonCommandLineOptionsValidationErrorPrefix, innerError));
-                }
-            }
-        }
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static async Task<ValidationResult> ValidateConfigurationAsync(
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>>.KeyCollection extensionsProviders,
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>>.KeyCollection systemProviders,
-        ICommandLineOptions commandLineOptions)
-    {
-        StringBuilder? stringBuilder = await ValidateConfigurationAsync(systemProviders, commandLineOptions, null).ConfigureAwait(false);
-        stringBuilder = await ValidateConfigurationAsync(extensionsProviders, commandLineOptions, stringBuilder).ConfigureAwait(false);
-
-        return stringBuilder?.Length > 0
-            ? ValidationResult.Invalid(stringBuilder.ToTrimmedString())
-            : ValidationResult.Valid();
-    }
-
-    private static async Task<StringBuilder?> ValidateConfigurationAsync(
-        Dictionary<ICommandLineOptionsProvider, IReadOnlyCollection<CommandLineOption>>.KeyCollection providers,
-        ICommandLineOptions commandLineOptions,
-        StringBuilder? stringBuilder)
-    {
-        foreach (ICommandLineOptionsProvider commandLineOptionsProvider in providers)
-        {
-            ValidationResult result = await commandLineOptionsProvider.ValidateCommandLineOptionsAsync(commandLineOptions).ConfigureAwait(false);
-            if (!result.IsValid)
-            {
-                stringBuilder ??= new();
-                stringBuilder.AppendLine(string.Format(CultureInfo.InvariantCulture, PlatformResources.CommandLineInvalidConfiguration, commandLineOptionsProvider.DisplayName, commandLineOptionsProvider.Uid, result.ErrorMessage));
-                stringBuilder.AppendLine();
-            }
-        }
-
-        return stringBuilder;
-    }
-
-    private static string ToTrimmedString(this StringBuilder stringBuilder)
-    {
-        // Trim trailing CR/LF characters directly from the StringBuilder to avoid extra allocations
-        while (stringBuilder.Length > 0 && stringBuilder[stringBuilder.Length - 1] is '\r' or '\n')
-        {
-            stringBuilder.Length--;
-        }
-
-        return stringBuilder.ToString();
-    }
-
-    private static ValidationResult AddCommandLine(CommandLineParseResult parseResult, ValidationResult result)
-        => result.IsValid ? result : InvalidWithCommandLine(parseResult, result.ErrorMessage);
-
-    private static ValidationResult InvalidWithCommandLine(CommandLineParseResult parseResult, string errorMessage)
-    {
-        errorMessage = parseResult.RedactError(errorMessage);
-        if (parseResult.CommandLine.Length == 0)
-        {
-            return ValidationResult.Invalid(errorMessage);
-        }
-
-        var stringBuilder = new StringBuilder(errorMessage);
-        stringBuilder.AppendLine();
-        stringBuilder.AppendFormat(CultureInfo.InvariantCulture, PlatformResources.CommandLineValidationCommandLine, parseResult.CommandLine);
-        return ValidationResult.Invalid(stringBuilder.ToString());
     }
 }
